Add kd-tree to ray tracing

+++ src/flat.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-float comparator(float i1, float i2, float i3, int param){ // 0 - знайти найменше, 1 - найбільше
+double comparator(double i1, double i2, double i3, int param){ // 0 - знайти найменше, 1 - найбільше
   if (param == 1){
     if (i1 > i2 && i1 > i3) return i1;
     else return i2 > i3 ? i2 : i3;
